bpymorse.set_viewport_perspective would cause a segfault
Segfaults still happen if the scene is started when the viewport is not fixed to the main camera
when updating the perspective to 'CAMERA'.
This fix makes sure that the camera is selected on startup but is not perfect.

If opened inedit mode and the camera is deselected before starting the game, then morse will still seg fault
due to trying to connect to a camera that has not been initialised.
Not sure if this is a MORSE issue or an UPBGE issue.