Adding a new type of component "feedbacksensor"
It will allow publisher and subscriber datastreams (IN and OUT)
to be attached to it and handle them accordingly.
This allows it to give feedback on the real position of an actuator that may
not follow the exact input command.
An example being a motor rpm having some windup time.