Added a new time strategy 'wall time'.
This is an unstable strategy in that it hasn't been extinsevley tested yet.
However it doesn't change any existing time funcitonality.

It aims to resolve a 'time offset' bug observed in the BestEffortStrategy that appears
when running simulations with complex .blend files. The .blend files take a while to load and that
time creates a constant error in the simulation time.