Backwards compatiability between 2.79 now nominally works.
we need two morse default autorun files as there are breaking changes in the .blend file. This is handelled with a blender version check.

Also Added the ability to use python virtual environments through an environment variable: MORSE_PYTHON_VENV
Setting this (with a valid virtual environment) will allow you to use non-standard versions of python (sadly a requirement of blender 3+ on ubuntu 20)

Tested with:
Ubuntu 18.04: blender 2.79a
UBuntu 20.04: blender 3.0.0